+++ examples/yoga-graphql/domain/user.ts
@@ -1,0 +1,8 @@
+export interface User {
+  id: string;
+  name: string;
+}
+
+export interface UserRepository {
+  findByIds(ids: string[]): Promise<User[]>;
+}