+++ packages/express/mod.ts
@@ -37,6 +37,7 @@ export type {
   Provider,
   ProviderType,
   Scope,
+  ScopeDisposeOptions,
   Token,
   ValidateOptions,
   ValidationIssue,