+++ packages/graphql/mod.ts
@@ -34,6 +34,7 @@ export type {
   Provider,
   ProviderType,
   Scope,
+  ScopeDisposeOptions,
   Token,
   ValidateOptions,
   ValidationIssue,